+++ ICaNer.Shared/DTOs/Product/EditeProductRequest.cs
@@ -7,8 +7,8 @@ namespace ICaNer.Shared.DTOs.Product
         public int Id { get; set; }
         public string Name { get; set; }
         public string Code { get; set; }
-        public double Price { get; set; }
-        public double Count { get; set; }
+        public long Price { get; set; }
+        public int Count { get; set; }
         public string Statuce { get; set; }
     }
 }
